+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -20,6 +20,7 @@ import { EstopTakeover } from '/app/organisms/EmergencyStop'
 import { FirmwareUpdateTakeover } from '/app/organisms/FirmwareUpdateModal/FirmwareUpdateTakeover'
 import { IncompatibleModuleTakeover } from '/app/organisms/IncompatibleModule'
 import { ModuleWizardFlows } from '/app/organisms/ModuleWizardFlows'
+import { OnDeviceLoginOverlayProvider } from '/app/organisms/ODD/OnDeviceLogin'
 import { QuickTransferFlow } from '/app/organisms/ODD/QuickTransferFlow'
 import { MaintenanceRunTakeover } from '/app/organisms/TakeoverModal'
 import { ToasterOven } from '/app/organisms/ToasterOven'
@@ -34,7 +35,6 @@ import { InstrumentDetail } from '/app/pages/ODD/InstrumentDetail'
 import { InstrumentsDashboard } from '/app/pages/ODD/InstrumentsDashboard'
 import { NameRobot } from '/app/pages/ODD/NameRobot'
 import { NetworkSetupMenu } from '/app/pages/ODD/NetworkSetupMenu'
-import { OnDeviceLoginOverlayProvider } from '/app/pages/ODD/OnDeviceLoginPage'
 import { ProtocolDashboard } from '/app/pages/ODD/ProtocolDashboard'
 import { ProtocolDetails } from '/app/pages/ODD/ProtocolDetails'
 import { ProtocolSetup } from '/app/pages/ODD/ProtocolSetup'
